K-56113 refactor tag and button component
