+++ src/component/tag/index.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const STag = styled.span`
+const Tag = styled.span`
   padding: 0.5rem;
 
   border-radius: 2px;
@@ -8,4 +8,4 @@ const STag = styled.span`
   color: ${props => props.theme.black};
 `;
 
-export default STag;
+export default Tag;
